Extract wallet_conf
